requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
multidict==6.0.4
numpy==1.24.3
opencv-python-headless==4.6.0.66
PyJWT==2.7.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
twilio==8.2.0
urllib3==2.0.2
yarl==1.9.2
